refactor(research): split cli support modules
