+++ gaia/cli/commands/research.py
@@ -2,31 +2,34 @@
 
 from __future__ import annotations
 
-import asyncio
 import json
-import os
-import re
-import shlex
-import subprocess
 import sys
-from concurrent.futures import ThreadPoolExecutor, as_completed
 from pathlib import Path
-from time import perf_counter, sleep
+from time import perf_counter
 from typing import Annotated, Any, cast
 
 import typer
 
-from gaia.cli.commands.add import (
-    LKMDependencyAddError,
-    add_lkm_chain_dependency,
-    add_lkm_claim_dependency,
-    add_lkm_paper_dependency,
-    add_local_package_dependency,
-    make_lkm_claim_ref,
-    make_lkm_paper_ref,
+from gaia.cli.commands.research_materialization import (
+    _materialize_landscape_sources_or_exit,
+    _materialize_lkm_papers_or_exit,
+)
+from gaia.cli.commands.research_providers import (
+    _json_object_from_llm_content,  # noqa: F401 - kept for CLI test compatibility
+    _load_research_env_files_or_exit,
+    _resolve_litellm_model,
+    _run_analysis_provider_command,
+    _run_analysis_provider_litellm,
+)
+from gaia.cli.commands.research_report_writing import _maybe_run_sectioned_report_writing
+from gaia.cli.commands.research_runtime import (
+    _emit_run_event,
+    _read_json_object_path,
+    _record_run_cli_trace,
+    _record_run_trace,
+    _update_run_state,
 )
 from gaia.cli.commands.search.lkm._indexes import DEFAULT_LKM_INDEX_ID
-from gaia.engine.packaging import GaiaPackagingError
 from gaia.engine.research import (
     AssessmentSchemaError,
     ProposalSchemaError,
@@ -35,7 +38,6 @@ from gaia.engine.research import (
     ResearchTargetError,
     ScanBatch,
     append_research_event,
-    attach_source_package_refs,
     build_assessment_from_analysis,
     build_assessment_from_landscapes,
     build_field_map_artifact,
@@ -46,9 +48,7 @@ from gaia.engine.research import (
     ensure_research_manifest,
     evaluate_research_stop,
     load_research_package,
-    materialize_landscape_source_package,
     render_final_research_report_markdown,
-    render_markdown_with_research_citations,
     render_research_artifact_markdown,
     research_contract,
     sync_assessment_artifact,
@@ -63,13 +63,7 @@ from gaia.engine.research.benchmark import (
     append_research_trace_step,
     write_research_benchmark_summary,
 )
-from gaia.engine.research.run import (
-    RUN_MODES,
-    ResearchRunStart,
-    append_run_event,
-    start_research_run,
-    write_run_state,
-)
+from gaia.engine.research.run import RUN_MODES, ResearchRunStart, start_research_run
 
 research_app = typer.Typer(
     name="research",
@@ -125,21 +119,6 @@ def _read_search_json(ref: str) -> tuple[dict[str, object], str]:
         typer.echo("Error: --search-json must contain a results array.", err=True)
         raise typer.Exit(2)
     return payload, label
-
-
-def _read_json_object_path(path: Path) -> dict[str, object]:
-    if not path.exists():
-        typer.echo(f"Error: file not found: {path}", err=True)
-        raise typer.Exit(2)
-    try:
-        payload = json.loads(path.read_text(encoding="utf-8"))
-    except json.JSONDecodeError as exc:
-        typer.echo(f"Error: file is not valid JSON: {path}: {exc}", err=True)
-        raise typer.Exit(2) from exc
-    if not isinstance(payload, dict):
-        typer.echo(f"Error: file must contain a JSON object: {path}", err=True)
-        raise typer.Exit(2)
-    return payload
 
 
 def _read_trace_records(trace_dir: Path) -> list[dict[str, object]]:
@@ -327,196 +306,6 @@ def _print_sync_summary(payload: dict[str, object]) -> None:
     focus_set = payload.get("focus_set")
     if isinstance(focus_set, str) and focus_set:
         typer.echo(f"focus_set: {focus_set}")
-
-
-def _materialize_landscape_sources_or_exit(
-    research_pkg: ResearchPackage,
-    landscape: dict[str, Any],
-    *,
-    landscape_artifact: Path,
-    enabled: bool,
-    artifact_only: bool,
-    dry_run: bool,
-) -> dict[str, object]:
-    if not enabled or artifact_only or dry_run:
-        return {
-            "materialize_sources_enabled": bool(enabled),
-            "source_package_materialization": False,
-            "source_packages_written": [],
-            "source_packages_added": [],
-        }
-
-    materialized = materialize_landscape_source_package(
-        research_pkg,
-        landscape,
-        landscape_artifact=landscape_artifact,
-    )
-    if materialized is None:
-        return {
-            "materialize_sources_enabled": True,
-            "source_package_materialization": False,
-            "source_packages_written": [],
-            "source_packages_added": [],
-        }
-
-    payload = materialized.to_payload()
-    try:
-        local_root = add_local_package_dependency(materialized.root, package_root=research_pkg.path)
-    except GaiaPackagingError as exc:
-        typer.echo(f"Error: failed to add generated source package: {exc}", err=True)
-        typer.echo(f"Generated source package: {materialized.root}", err=True)
-        raise typer.Exit(1) from exc
-
-    added_payload = dict(payload)
-    added_payload["path"] = str(local_root)
-    attach_source_package_refs(landscape, [materialized])
-    landscape_artifact.write_text(
-        json.dumps(landscape, indent=2, ensure_ascii=False) + "\n",
-        encoding="utf-8",
-    )
-    return {
-        "materialize_sources_enabled": True,
-        "source_package_materialization": True,
-        "source_packages_written": [payload],
-        "source_packages_added": [added_payload],
-    }
-
-
-def _lkm_materialized_payload(
-    materialized: Any,
-    *,
-    requested_ref: str | None = None,
-) -> dict[str, object]:
-    payload = {
-        "requested_source_ref": requested_ref or str(materialized.source_ref),
-        "source_ref": str(materialized.source_ref),
-        "path": str(materialized.root),
-        "package": str(materialized.dist_name),
-        "import_name": str(materialized.import_name),
-        "claim_count": int(materialized.claim_count),
-        "question_count": int(materialized.question_count),
-        "dependency_count": int(materialized.dependency_count),
-    }
-    chain_count = getattr(materialized, "chain_count", None)
-    if isinstance(chain_count, int):
-        payload["chain_count"] = chain_count
-    total_chains = getattr(materialized, "total_chains", None)
-    if isinstance(total_chains, int):
-        payload["total_chains"] = total_chains
-    return payload
-
-
-def _materialize_lkm_papers_or_exit(
-    research_pkg: ResearchPackage,
-    *,
-    paper_ids: list[str],
-    claim_ids: list[str],
-    chain_claim_ids: list[str],
-    lkm_index: str,
-    artifact_only: bool,
-    dry_run: bool,
-) -> dict[str, object]:
-    requests = [*paper_ids, *claim_ids, *chain_claim_ids]
-    if not requests or artifact_only or dry_run:
-        return {
-            "lkm_materialize_requests": requests,
-            "lkm_packages_materialized": [],
-            "lkm_chains_materialized": [],
-        }
-
-    materialized_packages = [
-        *_materialize_lkm_paper_refs(
-            research_pkg,
-            paper_ids=paper_ids,
-            lkm_index=lkm_index,
-        ),
-        *_materialize_lkm_claim_refs(
-            research_pkg,
-            claim_ids=claim_ids,
-            lkm_index=lkm_index,
-        ),
-    ]
-    materialized_chains = _materialize_lkm_chain_refs(
-        research_pkg,
-        claim_ids=chain_claim_ids,
-        lkm_index=lkm_index,
-    )
-    return {
-        "lkm_materialize_requests": requests,
-        "lkm_packages_materialized": materialized_packages,
-        "lkm_chains_materialized": materialized_chains,
-    }
-
-
-def _materialize_lkm_paper_refs(
-    research_pkg: ResearchPackage,
-    *,
-    paper_ids: list[str],
-    lkm_index: str,
-) -> list[dict[str, object]]:
-    materialized_packages: list[dict[str, object]] = []
-    for paper_id in paper_ids:
-        try:
-            ref = make_lkm_paper_ref(lkm_index, paper_id)
-            materialized = add_lkm_paper_dependency(ref, package_root=research_pkg.path)
-        except LKMDependencyAddError as exc:
-            typer.echo(f"Error: failed to add LKM paper package: {exc}", err=True)
-            if exc.materialized is not None:
-                typer.echo(f"Generated LKM package: {exc.materialized.root}", err=True)
-            raise typer.Exit(1) from exc
-        except GaiaPackagingError as exc:
-            typer.echo(f"Error: failed to materialize LKM paper {paper_id!r}: {exc}", err=True)
-            raise typer.Exit(1) from exc
-        materialized_packages.append(_lkm_materialized_payload(materialized, requested_ref=ref.ref))
-    return materialized_packages
-
-
-def _materialize_lkm_claim_refs(
-    research_pkg: ResearchPackage,
-    *,
-    claim_ids: list[str],
-    lkm_index: str,
-) -> list[dict[str, object]]:
-    materialized_packages: list[dict[str, object]] = []
-    for claim_id in claim_ids:
-        try:
-            ref = make_lkm_claim_ref(lkm_index, claim_id)
-            materialized = add_lkm_claim_dependency(ref, package_root=research_pkg.path)
-        except LKMDependencyAddError as exc:
-            typer.echo(f"Error: failed to add LKM claim backing package: {exc}", err=True)
-            if exc.materialized is not None:
-                typer.echo(f"Generated LKM package: {exc.materialized.root}", err=True)
-            raise typer.Exit(1) from exc
-        except GaiaPackagingError as exc:
-            typer.echo(f"Error: failed to materialize LKM claim {claim_id!r}: {exc}", err=True)
-            raise typer.Exit(1) from exc
-        materialized_packages.append(_lkm_materialized_payload(materialized, requested_ref=ref.ref))
-    return materialized_packages
-
-
-def _materialize_lkm_chain_refs(
-    research_pkg: ResearchPackage,
-    *,
-    claim_ids: list[str],
-    lkm_index: str,
-) -> list[dict[str, object]]:
-    materialized_chains: list[dict[str, object]] = []
-    for claim_id in claim_ids:
-        try:
-            ref = make_lkm_claim_ref(lkm_index, claim_id)
-            chain_materialized = add_lkm_chain_dependency(ref, package_root=research_pkg.path)
-        except LKMDependencyAddError as exc:
-            typer.echo(f"Error: failed to add LKM reasoning chain package: {exc}", err=True)
-            if exc.materialized is not None:
-                typer.echo(f"Generated LKM package: {exc.materialized.root}", err=True)
-            raise typer.Exit(1) from exc
-        except GaiaPackagingError as exc:
-            typer.echo(f"Error: failed to materialize LKM chain {claim_id!r}: {exc}", err=True)
-            raise typer.Exit(1) from exc
-        materialized_chains.append(
-            _lkm_materialized_payload(chain_materialized, requested_ref=ref.ref)
-        )
-    return materialized_chains
 
 
 @research_app.command("contract")
@@ -975,10 +764,6 @@ def run_command(
     typer.echo("phase: query_plan")
 
 
-def _run_state(run: ResearchRunStart) -> dict[str, Any]:
-    return _read_json_object_path(run.state_path)
-
-
 def _auto_plan_broad_queries_if_needed(
     research_pkg: ResearchPackage,
     run: ResearchRunStart,
@@ -1017,89 +802,6 @@ def _auto_plan_broad_queries_if_needed(
         json_stream=json_stream,
     )
     return _queries_from_query_plan(_read_json_object_ref(query_plan_json, label="query-plan JSON"))
-
-
-def _update_run_state(run: ResearchRunStart, updates: dict[str, Any]) -> dict[str, Any]:
-    state = _run_state(run)
-    state.update(updates)
-    write_run_state(run.state_path, state)
-    return state
-
-
-def _emit_run_event(
-    run: ResearchRunStart,
-    *,
-    event_type: str,
-    phase: str,
-    json_stream: bool,
-    payload: dict[str, Any] | None = None,
-) -> dict[str, Any]:
-    event = append_run_event(
-        run.events_path,
-        run_id=run.run_id,
-        event_type=event_type,
-        phase=phase,
-        payload=payload,
-    )
-    if json_stream:
-        typer.echo(json.dumps(event, ensure_ascii=False))
-    return event
-
-
-def _load_research_env_files_or_exit(env_file: list[str] | None) -> None:
-    refs = _research_env_file_refs(env_file)
-    for ref in refs:
-        path = Path(ref).expanduser()
-        if not path.exists():
-            typer.echo(f"Error: --env-file not found: {ref}", err=True)
-            raise typer.Exit(2)
-        if not path.is_file():
-            typer.echo(f"Error: --env-file is not a file: {ref}", err=True)
-            raise typer.Exit(2)
-        try:
-            assignments = _parse_research_env_file(path)
-        except ValueError as exc:
-            typer.echo(f"Error: invalid --env-file {ref}: {exc}", err=True)
-            raise typer.Exit(2) from exc
-        for key, value in assignments.items():
-            os.environ.setdefault(key, value)
-
-
-def _research_env_file_refs(env_file: list[str] | None) -> list[str]:
-    refs = [ref for ref in (env_file or []) if ref.strip()]
-    if refs:
-        return refs
-    configured = os.environ.get("GAIA_RESEARCH_ENV_FILE")
-    if configured:
-        return [ref for ref in configured.split(os.pathsep) if ref.strip()]
-    return []
-
-
-def _parse_research_env_file(path: Path) -> dict[str, str]:
-    assignments: dict[str, str] = {}
-    for line_number, raw_line in enumerate(path.read_text(encoding="utf-8").splitlines(), start=1):
-        line = raw_line.strip()
-        if not line or line.startswith("#"):
-            continue
-        if line.startswith("export "):
-            line = line.removeprefix("export ").strip()
-        if "=" not in line:
-            raise ValueError(f"line {line_number} must be KEY=VALUE")
-        key, value = line.split("=", 1)
-        key = key.strip()
-        if not re.fullmatch(r"[A-Za-z_][A-Za-z0-9_]*", key):
-            raise ValueError(f"line {line_number} has invalid key {key!r}")
-        assignments[key] = _parse_env_value(value.strip())
-    return assignments
-
-
-def _parse_env_value(value: str) -> str:
-    if len(value) >= 2 and value[0] == value[-1] and value[0] in {"'", '"'}:
-        return value[1:-1]
-    comment_index = value.find(" #")
-    if comment_index != -1:
-        value = value[:comment_index].rstrip()
-    return value
 
 
 def _write_run_checkpoint(
@@ -1145,67 +847,6 @@ def _write_run_checkpoint(
         phase=phase,
         json_stream=json_stream,
         payload={"pending_checkpoint": str(checkpoint_path)},
-    )
-
-
-def _record_run_cli_trace(
-    research_pkg: ResearchPackage,
-    run: ResearchRunStart,
-    *,
-    start: float,
-    name: str,
-    mode: str,
-    inputs: list[str] | None = None,
-    outputs: list[str] | None = None,
-    metrics: dict[str, object] | None = None,
-    model: str | None = None,
-    token_usage: dict[str, int] | None = None,
-    status: str = "ok",
-) -> None:
-    append_research_trace_step(
-        research_pkg,
-        run.run_dir / "trace",
-        name=name,
-        kind="cli",
-        mode=mode,
-        wall_seconds=perf_counter() - start,
-        inputs=inputs,
-        outputs=outputs,
-        metrics=metrics,
-        model=model,
-        token_usage=token_usage,
-        status=status,
-    )
-
-
-def _record_run_trace(
-    research_pkg: ResearchPackage,
-    run: ResearchRunStart,
-    *,
-    start: float,
-    name: str,
-    kind: str,
-    mode: str,
-    inputs: list[str] | None = None,
-    outputs: list[str] | None = None,
-    metrics: dict[str, object] | None = None,
-    model: str | None = None,
-    token_usage: dict[str, int] | None = None,
-    status: str = "ok",
-) -> None:
-    append_research_trace_step(
-        research_pkg,
-        run.run_dir / "trace",
-        name=name,
-        kind=kind,
-        mode=mode,
-        wall_seconds=perf_counter() - start,
-        inputs=inputs,
-        outputs=outputs,
-        metrics=metrics,
-        model=model,
-        token_usage=token_usage,
-        status=status,
     )
 
 
@@ -1318,609 +959,6 @@ def _run_lkm_knowledge_search(
         body["reasoning_only"] = True
     payload = run_request("POST", "/search", json_body=body, index_id=index)
     return normalize_lkm_knowledge(payload, query=query, kind="knowledge", index_id=index)
-
-
-def _run_analysis_provider_command(
-    research_pkg: ResearchPackage,
-    run: ResearchRunStart,
-    *,
-    phase: str,
-    command: str,
-    input_payload: dict[str, object],
-    output_name: str,
-    json_stream: bool,
-) -> str:
-    analysis_dir = run.run_dir / "analysis"
-    analysis_dir.mkdir(parents=True, exist_ok=True)
-    input_path = analysis_dir / f"{output_name}.input.json"
-    output_path = analysis_dir / f"{output_name}.output.json"
-    input_path.write_text(
-        json.dumps(input_payload, indent=2, ensure_ascii=False) + "\n",
-        encoding="utf-8",
-    )
-    _emit_run_event(
-        run,
-        event_type="provider.started",
-        phase=phase,
-        json_stream=json_stream,
-        payload={"provider": "command", "input": str(input_path), "output": str(output_path)},
-    )
-    args = shlex.split(command)
-    if not args:
-        typer.echo(f"Error: empty provider command for {phase}.", err=True)
-        raise typer.Exit(2)
-    env = {
-        **os.environ,
-        "GAIA_RESEARCH_PHASE": phase,
-        "GAIA_RESEARCH_INPUT": str(input_path),
-        "GAIA_RESEARCH_OUTPUT": str(output_path),
-        "GAIA_RESEARCH_RUN_DIR": str(run.run_dir),
-    }
-    start = perf_counter()
-    completed = subprocess.run(
-        args,
-        cwd=research_pkg.path,
-        env=env,
-        text=True,
-        capture_output=True,
-        check=False,
-    )
-    if completed.returncode != 0:
-        error = (completed.stderr or completed.stdout or "").strip()
-        _update_run_state(
-            run,
-            {
-                "status": "failed",
-                "phase": phase,
-                "error": error or f"provider command exited {completed.returncode}",
-            },
-        )
-        _emit_run_event(
-            run,
-            event_type="run.failed",
-            phase=phase,
-            json_stream=json_stream,
-            payload={"provider": "command", "returncode": completed.returncode, "error": error},
-        )
-        typer.echo(
-            f"Error: provider command failed for {phase} with exit code "
-            f"{completed.returncode}: {error}",
-            err=True,
-        )
-        raise typer.Exit(2)
-    if not output_path.exists():
-        typer.echo(
-            f"Error: provider command for {phase} did not write {output_path}.",
-            err=True,
-        )
-        raise typer.Exit(2)
-    _read_json_object_path(output_path)
-    _record_run_trace(
-        research_pkg,
-        run,
-        start=start,
-        name=f"provider.command.{phase}",
-        kind="llm",
-        mode="command",
-        inputs=[str(input_path)],
-        outputs=[str(output_path)],
-        metrics={
-            "provider": "command",
-            "phase": phase,
-            "returncode": completed.returncode,
-            "stdout_chars": len(completed.stdout or ""),
-            "stderr_chars": len(completed.stderr or ""),
-        },
-    )
-    _emit_run_event(
-        run,
-        event_type="provider.completed",
-        phase=phase,
-        json_stream=json_stream,
-        payload={"provider": "command", "input": str(input_path), "output": str(output_path)},
-    )
-    return str(output_path)
-
-
-def _run_analysis_provider_litellm(
-    research_pkg: ResearchPackage,
-    run: ResearchRunStart,
-    *,
-    phase: str,
-    model: str,
-    input_payload: dict[str, object],
-    output_name: str,
-    temperature: float,
-    timeout: float,
-    max_retries: int,
-    max_tokens: int | None,
-    json_stream: bool,
-) -> str:
-    rate_limit_retry_count = 2
-    rate_limit_retry_delay_seconds = 75.0
-    analysis_dir = run.run_dir / "analysis"
-    analysis_dir.mkdir(parents=True, exist_ok=True)
-    input_path = analysis_dir / f"{output_name}.input.json"
-    output_path = analysis_dir / f"{output_name}.output.json"
-    raw_path = analysis_dir / f"{output_name}.raw.txt"
-    hydrated_payload = _hydrate_analysis_provider_input(input_payload)
-    input_path.write_text(
-        json.dumps(hydrated_payload, indent=2, ensure_ascii=False) + "\n",
-        encoding="utf-8",
-    )
-    _emit_run_event(
-        run,
-        event_type="provider.started",
-        phase=phase,
-        json_stream=json_stream,
-        payload={
-            "provider": "litellm",
-            "model": model,
-            "input": str(input_path),
-            "output": str(output_path),
-        },
-    )
-    start = perf_counter()
-    response: object | None = None
-    retry_count = 0
-    retry_wait_seconds = 0.0
-    try:
-        while True:
-            try:
-                response = asyncio.run(
-                    _litellm_completion(
-                        model=model,
-                        phase=phase,
-                        input_payload=hydrated_payload,
-                        temperature=temperature,
-                        timeout=timeout,
-                        max_retries=max_retries,
-                        max_tokens=max_tokens,
-                    )
-                )
-                break
-            except Exception as exc:
-                if not _is_litellm_rate_limit_error(exc) or retry_count >= rate_limit_retry_count:
-                    raise
-                retry_count += 1
-                retry_wait_seconds += rate_limit_retry_delay_seconds
-                _emit_run_event(
-                    run,
-                    event_type="provider.retrying",
-                    phase=phase,
-                    json_stream=json_stream,
-                    payload={
-                        "provider": "litellm",
-                        "model": model,
-                        "reason": "rate_limit",
-                        "attempt": retry_count,
-                        "max_attempts": rate_limit_retry_count + 1,
-                        "wait_seconds": rate_limit_retry_delay_seconds,
-                        "error": str(exc),
-                    },
-                )
-                sleep(rate_limit_retry_delay_seconds)
-        content = _litellm_response_content(response)
-        raw_path.write_text(content, encoding="utf-8")
-        output_payload = _json_object_from_llm_content(content)
-    except Exception as exc:
-        usage = _litellm_usage_dict(response)
-        prompt_tokens = _int_metric(usage.get("prompt_tokens"))
-        completion_tokens = _int_metric(usage.get("completion_tokens"))
-        _record_run_trace(
-            research_pkg,
-            run,
-            start=start,
-            name=f"provider.litellm.{phase}",
-            kind="llm",
-            mode="litellm",
-            inputs=[str(input_path)],
-            outputs=[str(raw_path)] if raw_path.exists() else [],
-            metrics={
-                "provider": "litellm",
-                "phase": phase,
-                "model": model,
-                "error_type": type(exc).__name__,
-                "error": str(exc),
-                "raw_path": str(raw_path) if raw_path.exists() else None,
-                "prompt_tokens": prompt_tokens,
-                "completion_tokens": completion_tokens,
-                "total_tokens": _int_metric(usage.get("total_tokens")),
-                "request_id": _litellm_request_id(response),
-                "retry_count": retry_count,
-                "retry_wait_seconds": retry_wait_seconds,
-            },
-            model=model,
-            token_usage={
-                "input_tokens": prompt_tokens,
-                "output_tokens": completion_tokens,
-            },
-            status="failed",
-        )
-        _update_run_state(run, {"status": "failed", "phase": phase, "error": str(exc)})
-        _emit_run_event(
-            run,
-            event_type="run.failed",
-            phase=phase,
-            json_stream=json_stream,
-            payload={"provider": "litellm", "model": model, "error": str(exc)},
-        )
-        typer.echo(f"Error: LiteLLM provider failed for {phase}: {exc}", err=True)
-        raise typer.Exit(2) from exc
-    output_path.write_text(
-        json.dumps(output_payload, indent=2, ensure_ascii=False) + "\n",
-        encoding="utf-8",
-    )
-    usage = _litellm_usage_dict(response)
-    prompt_tokens = _int_metric(usage.get("prompt_tokens"))
-    completion_tokens = _int_metric(usage.get("completion_tokens"))
-    _record_run_trace(
-        research_pkg,
-        run,
-        start=start,
-        name=f"provider.litellm.{phase}",
-        kind="llm",
-        mode="litellm",
-        inputs=[str(input_path)],
-        outputs=[str(output_path)],
-        metrics={
-            "provider": "litellm",
-            "phase": phase,
-            "model": model,
-            "raw_path": str(raw_path),
-            "prompt_tokens": prompt_tokens,
-            "completion_tokens": completion_tokens,
-            "total_tokens": _int_metric(usage.get("total_tokens")),
-            "request_id": _litellm_request_id(response),
-            "retry_count": retry_count,
-            "retry_wait_seconds": retry_wait_seconds,
-        },
-        model=model,
-        token_usage={
-            "input_tokens": prompt_tokens,
-            "output_tokens": completion_tokens,
-        },
-    )
-    _emit_run_event(
-        run,
-        event_type="provider.completed",
-        phase=phase,
-        json_stream=json_stream,
-        payload={
-            "provider": "litellm",
-            "model": model,
-            "input": str(input_path),
-            "output": str(output_path),
-        },
-    )
-    return str(output_path)
-
-
-def _is_litellm_rate_limit_error(exc: Exception) -> bool:
-    text = str(exc).lower()
-    return "ratelimit" in text or "rate limit" in text or "tpm" in text
-
-
-async def _litellm_completion(
-    *,
-    model: str,
-    phase: str,
-    input_payload: dict[str, object],
-    temperature: float,
-    timeout: float,
-    max_retries: int,
-    max_tokens: int | None,
-) -> object:
-    os.environ.setdefault("LITELLM_LOCAL_MODEL_COST_MAP", "True")
-    import litellm
-
-    litellm.suppress_debug_info = True
-    litellm.disable_cost_calc = True
-    kwargs: dict[str, object] = {
-        "model": model,
-        "messages": _litellm_messages(phase=phase, input_payload=input_payload),
-        "temperature": temperature,
-        "timeout": timeout,
-        "max_retries": max_retries,
-        "response_format": {"type": "json_object"},
-    }
-    if max_tokens is not None:
-        kwargs["max_tokens"] = max_tokens
-    return await litellm.acompletion(**kwargs)
-
-
-def _litellm_messages(
-    *,
-    phase: str,
-    input_payload: dict[str, object],
-) -> list[dict[str, str]]:
-    return [
-        {
-            "role": "system",
-            "content": (
-                "You are Gaia's deterministic JSON compiler for research artifacts. "
-                "Return exactly one valid JSON object. The first non-whitespace "
-                "character must be `{` and the last must be `}`. Do not include "
-                "markdown, prose, XML, citations outside JSON, or code fences. Use "
-                "only source refs and ids present in the input artifact payloads."
-            ),
-        },
-        {
-            "role": "user",
-            "content": json.dumps(
-                {
-                    "phase": phase,
-                    "instruction": _litellm_phase_instruction(phase),
-                    "output_shape": _litellm_output_shape(phase),
-                    "validation_rules": [
-                        "Return a single JSON object, not an array or string.",
-                        "Do not add explanatory text before or after the JSON.",
-                        "Do not add decorative or prose-only keys outside the requested shape.",
-                        "Every source_refs entry must use ids present in artifact_payloads.",
-                        "If evidence is weak, encode uncertainty inside JSON fields.",
-                        "Prefer fewer high-quality grounded items over broad ungrounded output.",
-                        (
-                            "Keep strings compact; avoid LaTeX commands when Unicode "
-                            "text is available."
-                        ),
-                    ],
-                    "input": input_payload,
-                },
-                indent=2,
-                ensure_ascii=False,
-            ),
-        },
-    ]
-
-
-def _litellm_phase_instruction(phase: str) -> str:
-    if phase == "query_plan":
-        return (
-            "Generate 3-5 broad live-search queries for the topic. Cover distinct "
-            "evidence families likely to support an autonomous review map: "
-            "foundational theory, canonical models, methods/diagnostics, experiments "
-            "where relevant, and recent controversies. Do not assess evidence yet."
-        )
-    if phase == "field_map_analysis":
-        return (
-            "Induce a review-oriented field map from the broad landscape before "
-            "selecting narrow focuses. Build a taxonomy from primary retrieved "
-            "evidence: model families, methods, observables, theory constraints, "
-            "experimental systems, controversy axes, and coverage gaps. Recommend "
-            "only the highest-value live-search expansions needed for review coverage."
-        )
-    if phase == "focus_analysis":
-        return (
-            "Select 1-4 assessable research focuses from the landscapes and field map. "
-            "Each focus must be grounded in retrieved item ids, should fit into a "
-            "field-map bucket or controversy axis, and should be narrow enough for "
-            "immediate assessment."
-        )
-    if phase == "assess_analysis":
-        return (
-            "Assess only the selected focus against the selected evidence packet. "
-            "Produce at most 5 grounded candidate relations, limitations, and next "
-            "queries. Do not write the final review prose; later report phases will "
-            "write the article from this judgment. Emit at most 2 deferred "
-            "candidate obligations that directly follow from missing or conflicting "
-            "evidence. Set obligation.actionable=true only for a near-term blocking "
-            "task that should become an open inquiry obligation."
-        )
-    if phase == "report_plan":
-        return (
-            "Plan a scholarly evidence-review article from the field map, focus, "
-            "selected evidence, and assessment judgment. Produce a concise section "
-            "outline. Each section must have a stable id, title, purpose, and grounded "
-            "evidence refs. Do not reassess evidence; preserve the assessment judgment."
-        )
-    if phase == "report_section":
-        return (
-            "Write exactly one article section from the section plan, selected evidence, "
-            "section_evidence context, and assessment judgment. Use only refs and evidence "
-            "records present in section_evidence. If section_evidence.missing_refs is not "
-            "empty, explicitly describe that coverage gap instead of inventing details. Do "
-            "not change the assessment conclusion. Return section markdown, not a full article."
-        )
-    if phase == "report_stitch":
-        return (
-            "Stitch section drafts into one academic evidence-review report. Add title, "
-            "abstract, transitions, and conclusion. Remove repetition, keep citations as "
-            "inline source refs, and do not add workflow or benchmark commentary."
-        )
-    return "Produce the contract-shaped JSON for this research phase."
-
-
-def _litellm_output_shape(phase: str) -> dict[str, object]:
-    if phase == "query_plan":
-        return {
-            "required_top_level_keys": ["queries"],
-            "queries_item_shape": {"query": "search query text", "rationale": "why it helps"},
-        }
-    if phase == "field_map_analysis":
-        return {
-            "required_top_level_keys": [
-                "domain_thesis",
-                "buckets",
-                "controversy_axes",
-                "coverage_gaps",
-                "recommended_expansions",
-                "synthesis_notes",
-            ],
-            "buckets_item_keys": [
-                "id",
-                "title",
-                "role",
-                "required_for_review",
-                "coverage_status",
-                "evidence_refs",
-                "recommended_queries",
-            ],
-            "coverage_gaps_item_keys": ["kind", "description", "recommended_queries"],
-        }
-    if phase == "focus_analysis":
-        return {
-            "required_top_level_keys": ["focuses", "coverage_gaps", "notes"],
-            "focuses_item_keys": [
-                "id",
-                "kind",
-                "status",
-                "question",
-                "rationale",
-                "priority",
-                "readiness",
-                "scope",
-                "coverage",
-                "evidence_refs",
-                "suggested_queries",
-            ],
-        }
-    if phase == "assess_analysis":
-        return {
-            "required_top_level_keys": ["relations", "candidate_obligations"],
-            "relations_item_keys": [
-                "type",
-                "claim",
-                "rationale",
-                "epistemic_status",
-                "promotion_hint",
-                "source_refs",
-            ],
-            "optional_judgment_keys": ["limitations", "next_queries"],
-            "candidate_obligations_item_keys": [
-                "kind",
-                "content",
-                "source_refs",
-                "actionable",
-            ],
-        }
-    if phase == "report_plan":
-        return {
-            "required_top_level_keys": [
-                "title",
-                "abstract",
-                "thesis",
-                "sections",
-                "conclusion_prompt",
-            ],
-            "sections_item_keys": ["id", "title", "purpose", "evidence_refs"],
-        }
-    if phase == "report_section":
-        return {
-            "required_top_level_keys": ["section_id", "title", "markdown", "used_refs"],
-            "markdown": "Markdown for exactly one report section. Start with a level-2 heading.",
-        }
-    if phase == "report_stitch":
-        return {
-            "required_top_level_keys": ["markdown"],
-            "markdown": (
-                "Complete final report Markdown with title, abstract, body, and conclusion."
-            ),
-        }
-    return {"required_top_level_keys": []}
-
-
-def _hydrate_analysis_provider_input(payload: dict[str, object]) -> dict[str, object]:
-    hydrated = dict(payload)
-    artifact_payloads: list[dict[str, object]] = []
-    artifacts = payload.get("artifacts")
-    if isinstance(artifacts, list):
-        for artifact in artifacts:
-            if not isinstance(artifact, str):
-                continue
-            path = Path(artifact)
-            if path.exists():
-                artifact_payloads.append(
-                    {
-                        "path": artifact,
-                        "json": _read_json_object_path(path),
-                    }
-                )
-    hydrated["artifact_payloads"] = artifact_payloads
-    return hydrated
-
-
-def _json_object_from_llm_content(content: str) -> dict[str, object]:
-    text = content.strip()
-    fenced = re.fullmatch(r"```(?:json)?\s*(.*?)\s*```", text, flags=re.DOTALL)
-    if fenced:
-        text = fenced.group(1).strip()
-    try:
-        payload = json.loads(text)
-    except json.JSONDecodeError as exc:
-        repaired_text = _escape_invalid_json_backslashes(text)
-        try:
-            payload = json.loads(repaired_text)
-        except json.JSONDecodeError:
-            raise ValueError(f"LiteLLM response was not valid JSON: {exc}") from exc
-    if not isinstance(payload, dict):
-        raise ValueError("LiteLLM response must be a JSON object.")
-    return payload
-
-
-def _escape_invalid_json_backslashes(text: str) -> str:
-    """Preserve LaTeX-style backslashes inside otherwise valid JSON strings."""
-    return re.sub(r'\\(?!["\\/bfnrtu])', r"\\\\", text)
-
-
-def _litellm_response_content(response: object) -> str:
-    if isinstance(response, dict):
-        choices = response.get("choices")
-        if isinstance(choices, list) and choices:
-            first = choices[0]
-            if isinstance(first, dict):
-                message = first.get("message")
-                if isinstance(message, dict):
-                    content = message.get("content")
-                    return "" if content is None else str(content)
-    choices = getattr(response, "choices", None)
-    if choices:
-        message = getattr(choices[0], "message", None)
-        content = getattr(message, "content", None)
-        return "" if content is None else str(content)
-    return ""
-
-
-def _litellm_usage_dict(response: object) -> dict[str, object]:
-    usage = (
-        response.get("usage") if isinstance(response, dict) else getattr(response, "usage", None)
-    )
-    if usage is None:
-        return {}
-    if isinstance(usage, dict):
-        return dict(usage)
-    if hasattr(usage, "model_dump"):
-        dumped = usage.model_dump()
-        return dumped if isinstance(dumped, dict) else {}
-    try:
-        return dict(usage)
-    except Exception:
-        return {}
-
-
-def _litellm_request_id(response: object) -> str | None:
-    if isinstance(response, dict):
-        value = response.get("id") or response.get("request_id")
-        return str(value) if value else None
-    for attr in ("id", "request_id"):
-        value = getattr(response, attr, None)
-        if value:
-            return str(value)
-    return None
-
-
-def _int_metric(value: object) -> int:
-    return int(value) if isinstance(value, int | float) else 0
-
-
-def _resolve_litellm_model(model: str | None) -> str:
-    resolved = model or os.environ.get("GAIA_RESEARCH_LLM_MODEL")
-    if resolved and resolved.strip():
-        return resolved.strip()
-    typer.echo(
-        "Error: --analysis-provider litellm requires --model or GAIA_RESEARCH_LLM_MODEL.",
-        err=True,
-    )
-    raise typer.Exit(2)
 
 
 def _query_plan_provider_input(
@@ -2544,630 +1582,6 @@ def _run_evidence_select_and_deep_expand(
         payload={"artifact": str(selected_evidence_path), **materialized},
     )
     return selected_evidence_path, selected_evidence
-
-
-def _report_provider_input(
-    *,
-    phase: str,
-    topic: str,
-    language: str,
-    artifact_paths: list[Path],
-    focus: str,
-    extra: dict[str, object] | None = None,
-) -> dict[str, object]:
-    payload: dict[str, object] = {
-        "schema_version": 1,
-        "type": "gaia.research.report_request",
-        "phase": phase,
-        "topic": topic,
-        "language": language,
-        "focus": focus,
-        "artifacts": [str(path) for path in artifact_paths],
-    }
-    if extra:
-        payload.update(extra)
-    return payload
-
-
-def _report_ref_key(ref: object) -> tuple[str, str] | None:
-    if not isinstance(ref, dict):
-        return None
-    kind = ref.get("kind")
-    if not isinstance(kind, str) or not kind:
-        return None
-    ref_id = ref.get("id") or ref.get("paper_id") or ref.get("ref")
-    if isinstance(ref_id, int):
-        ref_id = str(ref_id)
-    if not isinstance(ref_id, str) or not ref_id:
-        return None
-    return kind, ref_id
-
-
-def _report_section_ref_keys(section: dict[str, object]) -> list[tuple[str, str]]:
-    refs = section.get("evidence_refs")
-    if not isinstance(refs, list):
-        return []
-    keys: list[tuple[str, str]] = []
-    for ref in refs:
-        key = _report_ref_key(ref)
-        if key is not None and key not in keys:
-            keys.append(key)
-    return keys
-
-
-def _item_report_ref_keys(item: dict[str, Any]) -> set[tuple[str, str]]:
-    keys: set[tuple[str, str]] = set()
-    kind = item.get("kind")
-    item_id = item.get("id")
-    if isinstance(kind, str) and isinstance(item_id, str) and item_id:
-        keys.add((kind, item_id))
-        if kind == "variable":
-            keys.add(("item", item_id))
-    stable_item_id = item.get("item_id")
-    if isinstance(stable_item_id, str) and stable_item_id:
-        keys.add(("item", stable_item_id))
-    source = item.get("source")
-    paper_id = source.get("paper_id") if isinstance(source, dict) else None
-    if isinstance(paper_id, str) and paper_id:
-        keys.add(("paper", paper_id))
-    package_ref_payload = item.get("package_ref")
-    package_ref = package_ref_payload.get("ref") if isinstance(package_ref_payload, dict) else None
-    if isinstance(package_ref, str) and package_ref:
-        keys.add(("package_ref", package_ref))
-    return keys
-
-
-def _paper_lead_report_ref_keys(lead: dict[str, Any]) -> set[tuple[str, str]]:
-    keys: set[tuple[str, str]] = set()
-    paper_id = lead.get("paper_id")
-    if isinstance(paper_id, str) and paper_id:
-        keys.add(("paper", paper_id))
-    variable_ids = lead.get("variable_ids")
-    if isinstance(variable_ids, list):
-        for variable_id in variable_ids:
-            if isinstance(variable_id, str) and variable_id:
-                keys.add(("variable", variable_id))
-    return keys
-
-
-def _citation_report_ref_keys(citation: dict[str, Any]) -> set[tuple[str, str]]:
-    keys: set[tuple[str, str]] = set()
-    paper_id = citation.get("paper_id")
-    if isinstance(paper_id, str) and paper_id:
-        keys.add(("paper", paper_id))
-    source_kind = citation.get("source_kind")
-    source_id = citation.get("source_id")
-    if isinstance(source_kind, str) and isinstance(source_id, str) and source_id:
-        keys.add((source_kind, source_id))
-    for field, kind in (("item_ids", "item"), ("variable_ids", "variable")):
-        values = citation.get(field)
-        if isinstance(values, list):
-            for value in values:
-                if isinstance(value, str) and value:
-                    keys.add((kind, value))
-    return keys
-
-
-def _relation_report_ref_keys(relation: dict[str, Any]) -> set[tuple[str, str]]:
-    keys: set[tuple[str, str]] = set()
-    source_refs = relation.get("source_refs")
-    if not isinstance(source_refs, list):
-        return keys
-    for ref in source_refs:
-        key = _report_ref_key(ref)
-        if key is not None:
-            keys.add(key)
-    return keys
-
-
-def _report_citation_from_item(item: dict[str, Any], citation_id: str) -> dict[str, Any]:
-    source = item.get("source")
-    source_payload = source if isinstance(source, dict) else {}
-    paper_id = source_payload.get("paper_id")
-    source_kind = (
-        "paper" if isinstance(paper_id, str) and paper_id else str(item.get("kind") or "item")
-    )
-    citation: dict[str, Any] = {
-        "id": citation_id,
-        "source_kind": source_kind,
-        "title": source_payload.get("paper_title") or item.get("title"),
-        "doi": source_payload.get("doi"),
-        "item_ids": [],
-        "variable_ids": [],
-    }
-    stable_item_id = item.get("item_id") or item.get("id")
-    if isinstance(stable_item_id, str) and stable_item_id:
-        citation["item_ids"].append(stable_item_id)
-    if item.get("kind") == "variable" and isinstance(item.get("id"), str):
-        citation["variable_ids"].append(str(item["id"]))
-    if isinstance(paper_id, str) and paper_id:
-        citation["paper_id"] = paper_id
-    elif isinstance(item.get("id"), str):
-        citation["source_id"] = item["id"]
-    return citation
-
-
-def _report_citations_from_items(items: list[dict[str, Any]]) -> list[dict[str, Any]]:
-    citations: list[dict[str, Any]] = []
-    by_key: dict[tuple[str, str], dict[str, Any]] = {}
-    for item in items:
-        source = item.get("source")
-        source_payload = source if isinstance(source, dict) else {}
-        paper_id = source_payload.get("paper_id")
-        if isinstance(paper_id, str) and paper_id:
-            key = ("paper", paper_id)
-        else:
-            item_id = item.get("id") or item.get("item_id")
-            if not isinstance(item_id, str) or not item_id:
-                continue
-            key = (str(item.get("kind") or "item"), item_id)
-        citation = by_key.get(key)
-        if citation is None:
-            citation = _report_citation_from_item(item, f"section_citation_{len(citations) + 1}")
-            by_key[key] = citation
-            citations.append(citation)
-            continue
-        stable_item_id = item.get("item_id") or item.get("id")
-        if isinstance(stable_item_id, str) and stable_item_id not in citation["item_ids"]:
-            citation["item_ids"].append(stable_item_id)
-        if (
-            item.get("kind") == "variable"
-            and isinstance(item.get("id"), str)
-            and item["id"] not in citation["variable_ids"]
-        ):
-            citation["variable_ids"].append(item["id"])
-    return citations
-
-
-def _append_unique_dict(
-    values: list[dict[str, Any]],
-    value: dict[str, Any],
-    *,
-    seen: set[str],
-) -> None:
-    key = json.dumps(value, ensure_ascii=False, sort_keys=True)
-    if key in seen:
-        return
-    seen.add(key)
-    values.append(dict(value))
-
-
-def _append_unique_keyed_dict(
-    values: list[dict[str, Any]],
-    value: dict[str, Any],
-    *,
-    key: str,
-    seen: set[str],
-) -> None:
-    if key in seen:
-        return
-    seen.add(key)
-    values.append(dict(value))
-
-
-def _item_report_dedupe_key(item: dict[str, Any]) -> str:
-    kind = item.get("kind") or "item"
-    source_id = item.get("id") or item.get("item_id")
-    if isinstance(source_id, str) and source_id:
-        return f"{kind}:{source_id}"
-    source = item.get("source")
-    paper_id = source.get("paper_id") if isinstance(source, dict) else None
-    return (
-        f"paper:{paper_id}"
-        if isinstance(paper_id, str) and paper_id
-        else json.dumps(
-            item,
-            ensure_ascii=False,
-            sort_keys=True,
-        )
-    )
-
-
-def _paper_lead_report_dedupe_key(lead: dict[str, Any]) -> str:
-    paper_id = lead.get("paper_id")
-    return (
-        str(paper_id)
-        if isinstance(paper_id, str) and paper_id
-        else json.dumps(
-            lead,
-            ensure_ascii=False,
-            sort_keys=True,
-        )
-    )
-
-
-def _artifact_evidence_packets(payload: dict[str, Any]) -> list[dict[str, Any]]:
-    packets: list[dict[str, Any]] = []
-    evidence_packet = payload.get("evidence_packet")
-    if isinstance(evidence_packet, dict):
-        packets.append(evidence_packet)
-    if payload.get("kind") in {"research_landscape", "selected_evidence"}:
-        packets.append(payload)
-    return packets
-
-
-def _collect_matching_packet_records(
-    packet: dict[str, Any],
-    *,
-    ref_keys: set[tuple[str, str]],
-    items: list[dict[str, Any]],
-    paper_leads: list[dict[str, Any]],
-    seen_items: set[str],
-    seen_leads: set[str],
-) -> set[tuple[str, str]]:
-    matched: set[tuple[str, str]] = set()
-    packet_items = packet.get("items")
-    if isinstance(packet_items, list):
-        for item in packet_items:
-            if not isinstance(item, dict):
-                continue
-            item_keys = _item_report_ref_keys(item)
-            if item_keys.isdisjoint(ref_keys):
-                continue
-            matched.update(item_keys.intersection(ref_keys))
-            _append_unique_keyed_dict(
-                items,
-                item,
-                key=_item_report_dedupe_key(item),
-                seen=seen_items,
-            )
-
-    packet_leads = packet.get("paper_leads")
-    if isinstance(packet_leads, list):
-        for lead in packet_leads:
-            if not isinstance(lead, dict):
-                continue
-            lead_keys = _paper_lead_report_ref_keys(lead)
-            if lead_keys.isdisjoint(ref_keys):
-                continue
-            matched.update(lead_keys.intersection(ref_keys))
-            _append_unique_keyed_dict(
-                paper_leads,
-                lead,
-                key=_paper_lead_report_dedupe_key(lead),
-                seen=seen_leads,
-            )
-    return matched
-
-
-def _collect_matching_relations(
-    payload: dict[str, Any],
-    *,
-    ref_keys: set[tuple[str, str]],
-    relations: list[dict[str, Any]],
-    seen_relations: set[str],
-) -> set[tuple[str, str]]:
-    matched: set[tuple[str, str]] = set()
-    payload_relations = payload.get("relations")
-    if not isinstance(payload_relations, list):
-        return matched
-    for relation in payload_relations:
-        if not isinstance(relation, dict):
-            continue
-        relation_keys = _relation_report_ref_keys(relation)
-        if relation_keys.isdisjoint(ref_keys):
-            continue
-        matched.update(relation_keys.intersection(ref_keys))
-        _append_unique_dict(relations, relation, seen=seen_relations)
-    return matched
-
-
-def _collect_matching_citations(
-    payload: dict[str, Any],
-    *,
-    ref_keys: set[tuple[str, str]],
-    citations: list[dict[str, Any]],
-    seen_citations: set[str],
-) -> set[tuple[str, str]]:
-    matched: set[tuple[str, str]] = set()
-    payload_citations = payload.get("citations")
-    if not isinstance(payload_citations, list):
-        return matched
-    for citation in payload_citations:
-        if not isinstance(citation, dict):
-            continue
-        citation_keys = _citation_report_ref_keys(citation)
-        if citation_keys.isdisjoint(ref_keys):
-            continue
-        matched.update(citation_keys.intersection(ref_keys))
-        _append_unique_dict(citations, citation, seen=seen_citations)
-    return matched
-
-
-def _collect_report_section_evidence(
-    section: dict[str, object],
-    *,
-    artifact_paths: list[Path],
-) -> dict[str, object]:
-    ref_keys = set(_report_section_ref_keys(section))
-    context: dict[str, object] = {
-        "refs": [
-            {"kind": kind, "id": ref_id} for kind, ref_id in _report_section_ref_keys(section)
-        ],
-        "items": [],
-        "paper_leads": [],
-        "relations": [],
-        "citations": [],
-        "missing_refs": [],
-    }
-    if not ref_keys:
-        return context
-
-    items = cast(list[dict[str, Any]], context["items"])
-    paper_leads = cast(list[dict[str, Any]], context["paper_leads"])
-    relations = cast(list[dict[str, Any]], context["relations"])
-    citations = cast(list[dict[str, Any]], context["citations"])
-    seen_items: set[str] = set()
-    seen_leads: set[str] = set()
-    seen_relations: set[str] = set()
-    seen_citations: set[str] = set()
-    matched_keys: set[tuple[str, str]] = set()
-
-    for path in artifact_paths:
-        if not path.exists():
-            continue
-        payload = _read_json_object_path(path)
-        for packet in _artifact_evidence_packets(payload):
-            matched_keys.update(
-                _collect_matching_packet_records(
-                    packet,
-                    ref_keys=ref_keys,
-                    items=items,
-                    paper_leads=paper_leads,
-                    seen_items=seen_items,
-                    seen_leads=seen_leads,
-                )
-            )
-        matched_keys.update(
-            _collect_matching_relations(
-                payload,
-                ref_keys=ref_keys,
-                relations=relations,
-                seen_relations=seen_relations,
-            )
-        )
-        matched_keys.update(
-            _collect_matching_citations(
-                payload,
-                ref_keys=ref_keys,
-                citations=citations,
-                seen_citations=seen_citations,
-            )
-        )
-
-    for citation in _report_citations_from_items(items):
-        _append_unique_dict(citations, citation, seen=seen_citations)
-
-    context["missing_refs"] = [
-        {"kind": kind, "id": ref_id}
-        for kind, ref_id in _report_section_ref_keys(section)
-        if (kind, ref_id) not in matched_keys
-    ]
-    return context
-
-
-def _sectioned_report_citations(section_contexts: list[dict[str, object]]) -> list[dict[str, Any]]:
-    citations: list[dict[str, Any]] = []
-    seen: set[str] = set()
-    for context in section_contexts:
-        context_citations = context.get("citations", [])
-        if not isinstance(context_citations, list):
-            continue
-        for citation in context_citations:
-            if isinstance(citation, dict):
-                _append_unique_dict(citations, citation, seen=seen)
-    for index, citation in enumerate(citations, start=1):
-        citation["id"] = f"report_citation_{index}"
-    return citations
-
-
-def _section_id(section: object, *, fallback: str) -> str:
-    if isinstance(section, dict):
-        raw = section.get("id")
-        if isinstance(raw, str) and raw:
-            return raw
-    return fallback
-
-
-def _safe_output_suffix(value: str) -> str:
-    safe = re.sub(r"[^A-Za-z0-9_.-]+", "_", value).strip("._-")
-    return safe[:80] or "section"
-
-
-def _sections_from_report_plan(plan: dict[str, object]) -> list[dict[str, object]]:
-    sections = plan.get("sections")
-    if not isinstance(sections, list):
-        return []
-    return [section for section in sections if isinstance(section, dict)]
-
-
-def _run_sectioned_report_writing(
-    research_pkg: ResearchPackage,
-    run: ResearchRunStart,
-    *,
-    topic: str,
-    language: str,
-    model: str,
-    focus: str,
-    field_map_path: Path | None,
-    focus_path: Path,
-    landscape_paths: list[Path],
-    selected_evidence_path: Path | None,
-    assessment_path: Path,
-    llm_temperature: float,
-    llm_timeout: float,
-    llm_max_retries: int,
-    llm_max_tokens: int | None,
-    report_section_concurrency: int,
-    json_stream: bool,
-) -> tuple[str | None, list[str]]:
-    artifact_paths = [
-        *([field_map_path] if field_map_path is not None else []),
-        focus_path,
-        *([selected_evidence_path] if selected_evidence_path is not None else []),
-        assessment_path,
-    ]
-    if selected_evidence_path is None:
-        return None, []
-
-    _update_run_state(run, {"phase": "report_plan"})
-    report_plan_json = _run_analysis_provider_litellm(
-        research_pkg,
-        run,
-        phase="report_plan",
-        model=model,
-        input_payload=_report_provider_input(
-            phase="report_plan",
-            topic=topic,
-            language=language,
-            artifact_paths=artifact_paths,
-            focus=focus,
-        ),
-        output_name="report_plan",
-        temperature=llm_temperature,
-        timeout=llm_timeout,
-        max_retries=llm_max_retries,
-        max_tokens=llm_max_tokens,
-        json_stream=json_stream,
-    )
-    report_plan = _read_json_object_ref(report_plan_json, label="report-plan JSON")
-    sections = _sections_from_report_plan(report_plan)
-    section_contexts = [
-        _collect_report_section_evidence(
-            section,
-            artifact_paths=[
-                *landscape_paths,
-                *artifact_paths,
-                Path(report_plan_json),
-            ],
-        )
-        for section in sections
-    ]
-
-    def write_section(index: int, section: dict[str, object], context: dict[str, object]) -> str:
-        section_id = _section_id(section, fallback=f"section_{index}")
-        output_name = f"report_section_{_safe_output_suffix(section_id)}"
-        return _run_analysis_provider_litellm(
-            research_pkg,
-            run,
-            phase="report_section",
-            model=model,
-            input_payload=_report_provider_input(
-                phase="report_section",
-                topic=topic,
-                language=language,
-                artifact_paths=[assessment_path, Path(report_plan_json)],
-                focus=focus,
-                extra={
-                    "section_id": section_id,
-                    "section": section,
-                    "section_evidence": context,
-                },
-            ),
-            output_name=output_name,
-            temperature=llm_temperature,
-            timeout=llm_timeout,
-            max_retries=llm_max_retries,
-            max_tokens=llm_max_tokens,
-            json_stream=json_stream,
-        )
-
-    concurrency = max(1, report_section_concurrency)
-    section_refs_by_index: dict[int, str] = {}
-    if concurrency == 1 or len(sections) <= 1:
-        for index, section in enumerate(sections, start=1):
-            section_refs_by_index[index] = write_section(
-                index,
-                section,
-                section_contexts[index - 1],
-            )
-    else:
-        with ThreadPoolExecutor(max_workers=min(concurrency, len(sections))) as executor:
-            futures = {
-                executor.submit(write_section, index, section, section_contexts[index - 1]): index
-                for index, section in enumerate(sections, start=1)
-            }
-            for future in as_completed(futures):
-                section_refs_by_index[futures[future]] = future.result()
-
-    section_refs = [section_refs_by_index[index] for index in range(1, len(sections) + 1)]
-    section_paths = [Path(section_ref) for section_ref in section_refs]
-
-    stitch_json = _run_analysis_provider_litellm(
-        research_pkg,
-        run,
-        phase="report_stitch",
-        model=model,
-        input_payload=_report_provider_input(
-            phase="report_stitch",
-            topic=topic,
-            language=language,
-            artifact_paths=[assessment_path, Path(report_plan_json), *section_paths],
-            focus=focus,
-        ),
-        output_name="report_stitch",
-        temperature=llm_temperature,
-        timeout=llm_timeout,
-        max_retries=llm_max_retries,
-        max_tokens=llm_max_tokens,
-        json_stream=json_stream,
-    )
-    stitch_payload = _read_json_object_ref(stitch_json, label="report-stitch JSON")
-    markdown = stitch_payload.get("markdown")
-    if not isinstance(markdown, str) or not markdown.strip():
-        return None, [report_plan_json, *section_refs, stitch_json]
-    rendered = render_markdown_with_research_citations(
-        markdown.strip(),
-        citations=_sectioned_report_citations(section_contexts),
-        language=language,
-    )
-    return rendered, [report_plan_json, *section_refs, stitch_json]
-
-
-def _maybe_run_sectioned_report_writing(
-    research_pkg: ResearchPackage,
-    run: ResearchRunStart,
-    *,
-    topic: str,
-    language: str,
-    analysis_provider: str,
-    research_mode: str,
-    model: str | None,
-    assess_model: str | None,
-    focus: str,
-    field_map_path: Path | None,
-    focus_path: Path,
-    landscape_paths: list[Path],
-    selected_evidence_path: Path | None,
-    assessment_path: Path,
-    llm_temperature: float,
-    llm_timeout: float,
-    llm_max_retries: int,
-    llm_max_tokens: int | None,
-    report_section_concurrency: int,
-    json_stream: bool,
-) -> tuple[str | None, list[str]]:
-    if analysis_provider != "litellm" or research_mode != "fast_package_native":
-        return None, []
-    return _run_sectioned_report_writing(
-        research_pkg,
-        run,
-        topic=topic,
-        language=language,
-        model=_resolve_litellm_model(assess_model or model),
-        focus=focus,
-        field_map_path=field_map_path,
-        focus_path=focus_path,
-        landscape_paths=landscape_paths,
-        selected_evidence_path=selected_evidence_path,
-        assessment_path=assessment_path,
-        llm_temperature=llm_temperature,
-        llm_timeout=llm_timeout,
-        llm_max_retries=llm_max_retries,
-        llm_max_tokens=llm_max_tokens,
-        report_section_concurrency=report_section_concurrency,
-        json_stream=json_stream,
-    )
 
 
 def _execute_file_provider_run(
